FormDesign - Simplified and related files.
This branch contains a greatly simplified approach to representing common SDC rules.  It supports many simple Boolean predicate rules, but strips out almost all other rule types, such as those using numerics ad objects of all types.  In this simplified model, these more complex rules must be implemented with programming code that references the SDC object framework.

The SDC object framework will be available in a separate GitHub project.  It is also available as part of the SDC XML Generator project.

This commit is not yet feature-complete.  When complete, it will form the part of the Connectathon2019 branch in this same project.