Develop (#8)
* FormDesign - Simplified and related files.

This branch contains a greatly simplified approach to representing common SDC rules.  It supports many simple Boolean predicate rules, but strips out almost all other rule types, such as those using numerics ad objects of all types.  In this simplified model, these more complex rules must be implemented with programming code that references the SDC object framework.

The SDC object framework will be available in a separate GitHub project.  It is also available as part of the SDC XML Generator project.

This commit is not yet feature-complete.  When complete, it will form the part of the Connectathon2019 branch in this same project.

* FormDesign - Simplified and related files.

This branch contains a greatly simplified approach to representing common SDC rules.  It supports many simple Boolean predicate rules, but strips out almost all other rule types, such as those using numerics ad objects of all types.  In this simplified model, these more complex rules must be implemented with programming code that references the SDC object framework.

The SDC object framework will be available in a separate GitHub project.  It is also available as part of the SDC XML Generator project.

This commit is not yet feature-complete.  When complete, it will form the part of the Connectathon2019 branch in this same project.

* Added more Boolean rules, a toggle (true/false/toggle) switch type, and some related enhancements

* Added more Boolean rules, a toggle (true/false/toggle) switch type, and some related enhancements

* Modified @associatedValue support for list items.  More changes to come for representing Boolean rules based on responses and @associatedValue

* Corrected SDC expressions after bad merge

* Stripped unused  qualifier attributes from SType data types.  Some changes/simplifications to declarative rules (AlternativesType).  Added SDCVendorTestTemplate.xml

* Multiple changes, including new Package list in SDCRetrieveForm.xsd

* Cleanup of "S" (simple) data types and Boolean rules.

* Sync up with GitHub

* Cleanup of some rules and annotations

* Changes mostly to DataTypes.xsd, ActionsType, RetrieveForm.xsd

* SDCRetrieveForm - use mult. FDFs, sub-Pkgs, URLs

Also, small improvements to SDCFormDesign.xsd

* New xslt; Unique constraint on @ID

THe xslt is still a work in progress - bugs are present.
The SDCFormDesign.xsd Scehma can now validate that all @ID values are unique.
Also, improved sample xml;